// src/generated/client/models/ThermalMonitorScheduleWeekDay.ts
/* tslint:disable */
/* eslint-disable */
// @ts-nocheck
/**
 * Vehicle Management Services (management)
 * Vehicle Management Services (management)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ThermalMonitorScheduleWeekDay = {
    Monday: 'MONDAY',
    Tuesday: 'TUESDAY',
    Wednesday: 'WEDNESDAY',
    Thursday: 'THURSDAY',
    Friday: 'FRIDAY',
    Saturday: 'SATURDAY',
    Sunday: 'SUNDAY'
} as const;
export type ThermalMonitorScheduleWeekDay = typeof ThermalMonitorScheduleWeekDay[keyof typeof ThermalMonitorScheduleWeekDay];


export function ThermalMonitorScheduleWeekDayFromJSON(json: any): ThermalMonitorScheduleWeekDay {
    return ThermalMonitorScheduleWeekDayFromJSONTyped(json, false);
}

export function ThermalMonitorScheduleWeekDayFromJSONTyped(json: any, ignoreDiscriminator: boolean): ThermalMonitorScheduleWeekDay {
    return json as ThermalMonitorScheduleWeekDay;
}

export function ThermalMonitorScheduleWeekDayToJSON(value?: ThermalMonitorScheduleWeekDay | null): any {
    return value as any;
}