// src/generated/client/models/ThermalMonitorType.ts
/* tslint:disable */
/* eslint-disable */
// @ts-nocheck
/**
 * Vehicle Management Services (management)
 * Vehicle Management Services (management)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Scheduled monitors are activated and deactivated based on a schedule. One-off monitors can run forever or be started and/or finished based on optional parameters.
 * @export
 */
export const ThermalMonitorType = {
    OneOff: 'ONE_OFF',
    Scheduled: 'SCHEDULED'
} as const;
export type ThermalMonitorType = typeof ThermalMonitorType[keyof typeof ThermalMonitorType];


export function ThermalMonitorTypeFromJSON(json: any): ThermalMonitorType {
    return ThermalMonitorTypeFromJSONTyped(json, false);
}

export function ThermalMonitorTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): ThermalMonitorType {
    return json as ThermalMonitorType;
}

export function ThermalMonitorTypeToJSON(value?: ThermalMonitorType | null): any {
    return value as any;
}